Add GitHub Actions workflow
Resolves #2

.github/workflow/ci.yaml uses matlab-actions/setup-matlab@v2 to
install the latest release of MATLAB on the runner and then runs the
tasks in buildfile.m using matlab-actions/run-build@v2. Note that the
tests require the Mapping Toolbox and Image Processing Toolbox which
we install using the `products` property in the setup-matlab task. Any
additional products that are needed by the tests in the future must
be added here. The names of the products can be found at
https://raw.githubusercontent.com/mathworks-ref-arch/matlab-dockerfile/main/mpm-input-files/R2024a/mpm_input_r2024a.txt

buildfile.m now uses buildplan(localfunctions) to create the "test"
task rather than using the default TestTask. This allows us to include
`toolbox/` on the path and ensure that the tests can see the toolbox
code.