Compile libtopotoolbox
buildfile.m adds `compileTask`, which finds the MATLAB-provided CMake
and compiles the libtopotoolbox bindings. If the MATLAB-provided CMake
does not exist, it prints a message and returns.

bindings/ contains the MEX C files that implement the bindings as well
as the CMakeLists.txt that sets up the build system.

bindings/tt_has_topotoolbox.c is a MEX binding for the has_topotoolbox
function from libtopotoolbox that always returns 1.

toolbox/internal/haslibtopotoolbox.m is a MATLAB function that wraps
the tt_has_topotoolbox MEX function. If it finds the MEX function it
runs it and returns the true output. If it does not find the MEX
function, it returns false.

The compile build task will build everything in a build/ directory in
the root of the repository. This can be deleted once buildtool is
finished.

.gitignore is updated to ignore both the build/ directory and the
topotoolbox/internal/mex directory that the MEX files get installed
to.

tests/testHasLibTopoToolbox.m adds a test that haslibtopotoolbox
returns true.

.github/workflows/ci.yaml adds the compile task to the CI
configuration. It also adds MATLAB Coder, which ensures that CMake is
available.