Implement snapshot tests
Snapshot testing verifies that changes that we make to our code do not
unexpectedly change the results. It compares snapshots of the output
from previously run tests to the output from the new code, and the
tests fail if the new output does not equal the old one. If we do need
make changes that will change the output of our functions, we first
will update the snapshots with the new output and then make the
changes. The procedure for doing this is outlined in a comment at the
top of testSnapshots.m. At the moment, all of this updating must be
done manually.

This code implements snapshot tests for topotoolbox3 and also uses
topotoolbox3 to generate snapshots that can be used elseswhere in our
ecosystem. Since topotoolbox3 already includes implementations of the
algorithms that we are adding to libtopotoolbox and elsewhere, it
makes sense to generate snapshots from the existing topotoolbox3
code. In cases such as `excesstopography`, where we know we want to
replace the MATLAB implementation with a new libtopotoolbox
implementation that provides different output, we can update the
snapshot repository with outputs from those algorithms when we want to
make those changes to topotoolbox3.

Snapshots are stored in the TopoToolbox/test_data repository, which is
added as a git submodule (`tests/snapshots`) here. The `.gitmodules`
file contains the metadata that git needs to track that repository.

Each dataset gets its own directory within that repository, and the
test harness finds all of the datasets and runs the snapshot tests on
them.

The snapshot tests currently run `fillsinks` and `identifyflats` with
their default arguments. `identifyflats` is applied to the filled DEM
so that it is more likely to discover flat regions. These are the
first two steps of the default flow routing algorithm implemented in
libtopotoolbox, so they can easily be turned into tests for
libtopotoolbox.

It is important to note that snapshot tests can be very strict: small
changes in the output will cause the tests to fail even if those
changes are not scientifically meaningful. It is possible to make the
tests less stringent by comparing output within certain tolerances,
for instance, or by verifying properties that the output should share
with the snapshot other than exact equality.

That said, I mostly want snapshots for testing libtopotoolbox against
the existing MATLAB implementations (see
https://github.com/TopoToolbox/libtopotoolbox/issues/122 for
example). While I could run and upload these snapshots on my own,
keeping the tests under version control in topotoolbox3 tracks the
provenance of the snapshots without needing to duplicate metadata, and
including them in the automated tests ensures that the snapshots do
indeed reflect the current and expected behavior of TopoToolbox.

If it does become burdensome to update the snapshots after making
small changes to the MATLAB implementations, we can exclude these
tests from the checks that run for pull requests and merges on GitHub.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>