Implement fillsinks MEX function and use it within GRIDobj.fillsinks
Resolves #19

This commit adds `bindings/tt_fillsinks.c`, which implements a MEX
function that calls libtopotoolbox's `fillsinks`. It allocates memory
for the output and intermediate arrays within the MEX function and
passes pointers to those arrays to libtopotoolbox. Allocation failures
will terminate the MEX function and return control to MATLAB.

bindings/CMakeLists.txt adds the tt_fillsinks MEX target to the
compilation pipeline. This currently must be done manually for each
MEX file we want to compile.

This requires libtopotoolbox release 2024-W43. CMakeLists.txt
currently uses the main branch of libtopotoolbox. This commit should
not be merged until 2024-W43 is released (2024-10-21) and the git tag
is set appropriately.

toolbox/@GRIDobj/fillsinks.m is updated to use `tt_fillsinks` when
passed a single argument. We could probably use libtopotoolbox code
for the maximum depth and minima imposition as well, but the current
`fillsinks` function always does the preprocessing required for the
full sink filling algorithm before calling our `reconstruct`
function. `reconstruct` is not exported by libtopotoolbox at the
moment, though it could be. If libtopotoolbox is not installed
properly, this falls back on the former version using `imreconstruct`.

The existing snapshot tests should catch any problems with
`tt_fillsinks`, so no additional tests are added.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>