Obtain snapshots from the snapshot_data repository release
This workflow runs the snapshot tests using the data that are included
in the releases of our snapshot_data repository. The snapshot_data
repository is checked out and the snapshots are downloaded, then the
SHA256 checksums that are included in the snapshot_data repository are
compared to the data to ensure that everything matches. This check is
skipped on Windows, because of that platforms lack of a suitable
shasum equivalent.

GitHub Actions' cache is used to save time and bandwidth when
downloading the snapshot data. The caches are versioned using keys
based on the release number of the snapshot_data repository. When the
snapshots need to be updated, bump that version number to the
appropriate one.

The snapshots_data repository keeps everything in a `data/`
subdirectory, so the snapshot tests are updated to find data there.