Implement binding for libtopotoolbox identifyflats
Resolves #20

Adds a new MEX file, tt_identifyflats in bindings/ and the appropriate
CMakeLists.txt declaration. This depends on recent changes to
libtopotoolbox's `identifyflats` that handle NaNs correctly. The
version of libtopotoolbox using in bindings/CMakeLists.txt is set to
`main` for testing, but it should be fixed at `2024-W44` once that
release is made. This PR should not be merged before that release is
available.

@GRIDobj/identifyflats.m is updated to use tt_identifyflats for
identifying flats and sills. Closed basins are not handled by
libtopotoolbox's identifyflats at the moment, so they are processed as
they were before, using `imregionalmin`. If libtopotoolbox is not
available the implementation falls back to the previous Image
Processing Toolbox functions.