Add a snapshot test for the auxiliary topography
This adds a snapshot test that calls `createAuxiliaryTopo` on the test
DEM and either saves the resulting auxiliary topography or compares it
to the saved snapshot. The TopoToolbox/snapshot_data repository will
need to be updated manually if this PR is merged.

`createAuxiliaryTopo` was previously a private function under
`@FLOWobj/`, which means it was not accessible from outside of
`FLOWobj` methods and could not be tested on its own. It is useful to
have the snapshot of the auxiliary topography, because it can help
distinguish problems in flow routing over the auxiliary topography
from problems in generating the auxiliary topography itself, which I
think are important to rule out for TopoToolbox/libtopotoolbox#122.

This PR moves `createAuxiliaryTopo.m` and `getPreSillPixels.m` to
`toolbox/internal` so they can be accessed from the test. They could
also be conceivably be methods on `@GRIDobj`, but if we do not want
users to rely on them, we should probably keep them in `internal`.

It might also make sense to move `getPreSillPixels` into
`identifyflats`. libtopotoolbox returns the presills along with the
flats and sills in its `identifyflats`. However, this might complicate
how sills are identified for internal drainage basins, so I have left
that for future work, especially once we are able to label closed
basins in libtopotoolbox's
`identifyflats` (TopoToolbox/libtopotoolbox#141).