Add tt_gwdt_computecosts MEX function to CMake build system
To compile a libtopotoolbox MEX binding for use with topotoolbox3, one
must declare the MEX object in CMake with

```
matlab_add_mex(
  NAME tt_gwdt_computecosts
  MODULE
  SRC tt_gwdt_computecosts.c
  R2018a
  LINK_TO topotoolbox
)
```

The `NAME` argument will become the name of the compiled MEX
file (i.e. `tt_gwdt_computecosts.mexa64`) and therefore the name of
the MATLAB function used to call the MEX function. The `SRC` argument
is the name of the C file that implements the MEX function that we
added in the previous commit. `LINK_TO` should include `topotoolbox`,
which causes CMake to link to the static libtopotoolbox library that
it compiles.

Finally, one must also add the `NAME` of the declared MEX file to the
`install` `TARGETS` option. This ensures that the compiled MEX file
gets put in the right place when we compile it.