Validate the function interface
I have yet to develop a complete set of interface checks for the
libtopotoolbox bindings, and I usually just check that there are the
right number of input and output arrays. One could do more such as
checking the types or the sizes of the arrays. Since the bindings are
not intended to be used directly, we can probably get away with less
validation in the MEX function. Developing better interface validation
is worth thinking about as we go forward: perhaps the validation code
is a good way to document the interface more thoroughly.