Extract input and create output arrays
This uses the mxGet* functions to obtain pointers to the various input
arrays and the mxCreateNumericMatrix function to allocate the single
output array. The output array is managed by MATLAB and does not need
to be freed.

This will still not compile because the intermediate arrays have not
been created.