Create and destroy the connected components array
This intermediate array is used during gwdt_computecosts to store the
connected components of flat pixels. We allocate it using mxMalloc and
destroy it once we are finished with mxFree. Technically the mxFree is
not necessary because the MATLAB memory manager automatically frees
memory allocated with mxMalloc.

We could also create a numeric matrix to store the intermediate array,
which I have done in other bindings. This approach is arguably better
because it does not require coercing the ptrdiff_t type of the
intermediate array to a MATLAB type.

Note that the dimensions and size of the allocated array are checked
for signed integer overflows before they are multiplied together. It
is highly unlikely that these will overflow because an array with that
size would require way too much memory to exist. These checks have
therefore been omitted in previous bindings, but are included here for
completeness.

The function should now compile.