Bump libtopotoolbox version
Resolves #21

This commit contains the changes in both
TopoToolbox/libtopotoolbox#155 and TopoToolbox/libtopotoolbox#147, which
change the value of the constants `tweight` and `CarveMinVal` in
`gwdt_computecosts` to match the MATLAB implementation. It is
necessary to use this version to ensure that the snapshot tests pass
when using the libtopotoolbox implementation of
`gwdt_computecosts`. This commit will also be included in version
2024-W50 of libtopotoolbox upon its release, at which point the hash
can be bumped to the appropriate version number.