Reconfigure the datasets so that tests are labeled by site names
Previously, the tests were labeled by the paths to the datasets, which
sometimes got truncated, making it hard to see which dataset was used
for which test. By setting the available_datasets variable with the
directory names, each test is labeled by the site
name (i.e. "bigtujunga_30m", "tibet", etc.), and the test results are
more legible.