Export benchmark data and provide a way to plot it
This commit implements a system for maintaining a historical record of
benchmarks and one plot that will be particularly useful for
libtopotoolbox performance analysis going forward.

The benchmark task in buildfile.m is updated with some data wrangling
code that summarizes the timing results, extracts the test dataset,
function and libtopotoolbox flags for each snapshot run and appends
some information about the commit to the dataset. It then saves this
table as a CSV file (test/results/benchmark.csv). If that benchmark
results file already exists, it will be read, and the new benchmark
run will be appended to it.

Variations in the performance characteristics of different machines
make the benchmark results hard to compare when run on different
machines. The benchmark results are best used locally by individual
developers to evaluate the effects on performance of potential code
changes.

The benchmark is still not hooked up to CI, and the benchmark results
file should not be committed to the repository. It is possible to do
this, but it requires some more thought about what should be tracked
and how to store and publish the results.

A function, benchmarkComparisonPlot is added to the internal
directory. This function takes a table representing the results of the
benchmark (obtained with
`readtable("tests/results/benchmark.csv","TextType","String")`), and
produces a plot inspired by MATLAB's builtin `comparisonPlot`
function. It plots the benchmark timings with libtopotoolbox against
those without. Each of the measurements in the table for seven
snapshot datasets is plotted as a line, and the most recent test is
plotted as a colored circle. This plot can show changes in the
performance of either the MATLAB or the libtopotoolbox implementation
over time, and it shows how performance scales with DEM size.