new function propagatevaluesupstream and minor changes
This commit includes a few minor changes and a new function called propagatevaluesupstream. The function generalizes numerous computations that are performed by traversing flow networks in upstream direction, assigning values obtained at some nodes to their contributing nodes. The function simplifies calls to vertdistance2stream or mapfromnal, and potentially also could be used in the function drainagebasins. Potentially, it could also be used as a wrapper function for a C-function.