Reorganize documentation
This commit introduces a new directory toolbox/docs. This directory
contains the raw documentation files and scripts that will be compiled
into the toolbox documentation and possibly documentation for our
website. The guidance from Mathworks on documentation structure for
toolboxes is unclear, but this seems like a relatively good way to
organize things. We may also be able to move/remove the toolbox/help
directory and compile PDF documentation in the same way.

The mlx files are moved from toolbox/html/mlxfiles to
toolbox/docs/mlxfiles.

toolbox/info.xml is updated to refer to the new location of the help files.

publishtthelp2html.m updated to use the `export` function. It will put
the exported files in toolbox/docs/html.

The mlx files are unevaluated, and they should be left
unevaluated. The publishing script reevaluates them when building the
docs. Doing it this way tracks changes to the content of the notebooks
not the outputs, which might change due to things outside of our
control such as changes to MATLAB's plots. This is what we do with our
Jupyter notebook documentation in pytopotoolbox.

Unfortunately, the `export` function with the `Run=true` option can be
very flaky. Sometimes it hangs forever while reporting graphics
timeouts. It seems to run best when starting MATLAB fresh, so it might
not be such a problem to run in CI. More testing is needed before we
can build the docs adequately in CI.

toolbox/docs/helptoc.xml is the help table of contents that we
need. This is written by hand, not automatically generated, so it
should be committed into the repository. The publishing function
copies it into the toolbox/docs/html directory as needed.

The generated html files are all deleted, and their new location is
ignored in .gitignore. We should not commit autogenerated
documentation files to the git repository. Commit the .mlx files that
are used to generate the documentation instead.

A new `doc` task is added to buildfile.m. This runs our publishing
function to generate new documentation.