Only use the MATLAB CI cache on ubuntu and macos
It seems like when we restore the MATLAB installation from the cache,
it cannot be found by CMake, which causes the libtopotoolbox
compilation step and then haslibtopotoolbox test to fail. This commit
sets `cache: true` only on ubuntu and macos runners, requiring a new
install of MATLAB on Windows. This is intended as a temporary measure
to make sure we avoid CI errors. We might be able to help CMake find
the restored cache, but that requires some more detailed
investigation.