Add an option to skip creating figures in tests
The figures created in testFLOWobj may be useful when running the
tests interactively, but on CI we don't check the figures and it slows
things down. This change introduces a useGraphics argument to the test
task so that

buildtool test(useGraphics=false)

will not try to create the figures in testFLOWobj. The useGraphics
argument sets a ClassSetupParameter called graphics, which is then
used to skip the figure creation code. The graphics
ClassSetupParameter could also be used in other tests including other
classes if desired.

I have also turned the graphics off on CI by default.