Modify reaches above knickpoints in lowerenv/allpred
The existing update step in allpred sets the source node if the target
node is set and if the target node is not a knickpoint. This means
that if the starting node is directly above a knickpoint, it will get
toggled off, and then nodes in the reach above that starting node will
not get set. lowerenv will then not check that reach, and you can get
nonconvex segments above the knickpoints.

This change moves the `~kn(ixc(r))` check inside the parenthesis, so
the starting node cannot be unset by the presence of a
knickpoint. Reaches above knickpoints are now checked by lowerenv, and
the stream profiles are always convex, except at knickpoints.

This change makes the `lowerenv_convex` tests in `testKnickpoints.m`
pass.