Bind libtopotoolbox's lowerenv
This PR adds the uselibtt option to lowerenv and the
knickpointfinder. libtopotoolbox's lowerenv can be somewhat faster (2x
or so), but is numerically slightly different from MATLAB's
lowerenv. This will cause slight discrepancies in the identified
knickpoints, particularly at low tolerances.

The test is modified slightly to account for the numerical issues. The
minimum acceptable upslope gradient difference is changed from -1e-6
to -1e-2, and the convexity test is only applied on projected
DEMs. These tests now pass. 0.01 is a fairly large discrepancy for a
gradient, but note that the estimated profiles using libtopotoolbox
and MATLAB implementations differ by only a centimeter at most.