Minor changes and new function
This implements some minor changes to GRIDobj/distance and createRasterFromFile.
Also, the function greedylabel is added. The function uses a greedy approach to reduce the number of labels in a grid with labeled regions while ensuring that no adjacent regions share the same label. This is a simplified solution to the map-coloring problem.