Add CITATION.cff file and Zenodo DOI badge to README
The CITATION.cff file was automatically generated by Zenodo, then
lightly edited. Because we currently automatically release `latest` on
every commit to main, we won't be able to update this with the DOI for
the `latest` release. Therefore, the DOI given in the CITATION.cff
file is the "Concept" DOI provided by Zenodo, which will always
redirect to the latest release.