Rename the automatic releases
Creating and deleting the latest tag does not play well with Zenodo,
so we will instead create releases numbered

release-YYYYMMDD-N

where N is a number that increases for every release, so we can have
more than one per day. There is a little bit of effort required to
figure out what that number should be. Otherwise everything is the
same. The release will show up under the Release section on the
landing page with the latest tag, so it should not be too onerous to
find for users, even if the release number is a little intimidating.